Split server.go into multiple files based on their category
