+++ http/router.go
@@ -26,7 +26,7 @@ func SetUpRouter(router *gin.Engine, server *server.Server) {
 			protected.GET("/users/:id", handleGetUser(server))
 			protected.POST("/users", handleAddUser(server))
 			protected.POST("/sessions", handleAddSession(server))
-			protected.POST("/sessions/:id/attendance-form", handleCreateSessionForm(server))
+			protected.POST("/sessions/:id/attendance-form", handleCreateAttendanceForm(server))
 			protected.POST("/sessions/:id/attendance", handleApplyAttendance(server))
 		}
 	}
