+++ server/attendance.go
@@ -1,0 +1,66 @@
+package server
+
+import (
+	"errors"
+	"fmt"
+	"rush/session"
+	"sort"
+	"time"
+)
+
+func (s *Server) CreateAttendanceForm(sessionId string) (string, error) {
+	dbSession, err := s.sessionRepo.Get(sessionId)
+	if err != nil {
+		return "", newNotFoundError(fmt.Errorf("failed to get session: %w", err))
+	}
+	if dbSession.IsClosed {
+		return "", newBadRequestError(errors.New("session is already closed"))
+	}
+
+	users, err := s.userRepo.GetAll()
+	if err != nil {
+		return "", newInternalServerError(fmt.Errorf("failed to get users: %w", err))
+	}
+
+	sort.Slice(users, func(i, j int) bool {
+		if users[i].Generation != users[j].Generation {
+			return users[i].Generation > users[j].Generation
+		}
+		return users[i].Name < users[j].Name
+	})
+
+	if dbSession.GoogleFormUri != "" {
+		return "", newBadRequestError(fmt.Errorf("form already exists: URI is %s", dbSession.GoogleFormUri))
+	}
+
+	formTitle := fmt.Sprintf("[출석] %s", dbSession.Name)
+	startsAt := dbSession.StartsAt.In(s.formTimeLocation)
+	expiresAt := startsAt.Add(-time.Second)
+	formDescription := fmt.Sprintf(`%s을(를) 위한 출석용 구글폼입니다.
+폼 마감 시각은 %s입니다. %s 이후 요청은 무시됩니다.`, dbSession.Name, expiresAt.Format("2006-01-02 15:04:05"), startsAt.Format("2006-01-02 15:04:05"))
+
+	attendanceForm, err := s.attendanceFormHandler.GenerateForm(formTitle, formDescription, users)
+	if err != nil {
+		return "", newInternalServerError(fmt.Errorf("failed to generate form: %w", err))
+	}
+
+	_, err = s.sessionRepo.Update(sessionId, &session.UpdateForm{
+		GoogleFormId: &attendanceForm.Id, GoogleFormUri: &attendanceForm.Uri, ReturnUpdatedSession: false})
+	if err != nil {
+		return "", newInternalServerError(fmt.Errorf("failed to update session: %w", err))
+	}
+
+	return attendanceForm.Uri, nil
+}
+
+func (s *Server) GetAttendanceByUserId(userId string) ([]Attendance, error) {
+	attendances, err := s.attendanceRepo.FindByUserId(userId)
+	if err != nil {
+		return nil, newInternalServerError(fmt.Errorf("failed to find attendance by user ID: %w", err))
+	}
+	converted := []Attendance{}
+	for _, attendance := range attendances {
+		converted = append(converted, *fromAttendance(&attendance))
+	}
+	return converted, nil
+}
